QA (Radar / Hammer) found DTMs couldn't be adjusted in the TAG/Indirect phase: the Flip Mount and Rotate Turret buttons only existed in FiringDisplay, never in TargetingPhaseDisplay (which drives both the TARGETING and OFFBOARD/TAG phases). The right-click MapMenu worked, but "there's no way to adjust DTM due to the lack of button." Hammer confirmed: "That is a bug — they should be adjusted in the phase you'd fire TAG."
The fix (consistent button usage, per the request)

Rather than duplicate ~150 lines into the targeting display (which reviewers flag), I hoisted the shared DTM/turret controls into the common base AttackPhaseDisplay:

- Moved flipDirectionalMount, rotateSelectedMount, updateFlipMount, updateRotateTurret, pendingDirectionalMountFacings, findMekTurretItem into the base.
- Added four overridable no-op hooks (clearAttacks, refreshTargetAfterMountChange, setFlipMountEnabled, setRotateTurretEnabled) so the display-specific bits stay in each subclass without forcing the 3 unrelated attack displays (Physical, Infantry, PreEnd) to implement anything.
- FiringDisplay now inherits the shared logic (just keeps its own button setters + one hook); removed two now-unused imports.
- TargetingPhaseDisplay gets the two buttons: enum entries, action handlers, button setters, and enable-updates wired into updateTarget()/selectEntity()/disableButtons(), plus resource strings.

Two correctness details I confirmed/handled:
- Server applies it in the TAG phase — both TARGETING and OFFBOARD call resolveAllButWeaponAttacks() before weapon resolution, so the queued DirectionalMountFahase flip is applied server-side (the rotatedialog already sends immediately).
- Twist no longer cancels the flip in targeting — since Flip Mount now exists there, I made the targeting-phase torso twist preserve pending mount facings and thewist), matching the fix already in the firing phase.
