+++ megamek/resources/megamek/client/messages.properties
@@ -4385,6 +4385,10 @@ TargetingPhaseDisplay.DontFireDialog.title=Don't fire?
 TargetingPhaseDisplay.Fire=Fire
 TargetingPhaseDisplay.fireFire=Fire
 TargetingPhaseDisplay.fireFlipArms=Flip Arms
+TargetingPhaseDisplay.fireFlipMount=Flip Mount
+TargetingPhaseDisplay.fireFlipMount.tooltip=Flip the selected weapon's Directional Torso Mount between the front and rear arc (BMM p.83). The chosen arc persists until changed again.
+TargetingPhaseDisplay.fireRotateTurret=Rotate Turret
+TargetingPhaseDisplay.fireRotateTurret.tooltip=Rotate the selected weapon's turret or Directional Torso Mount to a new facing. Opens a facing picker; the chosen facing persists until changed again.
 TargetingPhaseDisplay.its_others_turn=It''s {0}''s turn to declare indirect attacks.
 TargetingPhaseDisplay.its_your_turn=It's your turn to declare indirect attacks.
 TargetingPhaseDisplay.its_your_tag_turn=It's your turn to declare offboard / TAG attacks.
