+++ megamek/src/megamek/client/ui/panels/phaseDisplay/FiringDisplay.java
@@ -47,7 +47,6 @@ import megamek.client.ui.clientGUI.ClientGUI;
 import megamek.client.ui.clientGUI.boardview.BoardView;
 import megamek.client.ui.clientGUI.boardview.IBoardView;
 import megamek.client.ui.clientGUI.boardview.overlay.ToastLevel;
-import megamek.client.ui.dialogs.TurretFacingDialog;
 import megamek.client.ui.dialogs.phaseDisplay.BombPayloadDialog;
 import megamek.client.ui.dialogs.phaseDisplay.SuicideImplantsDialog;
 import megamek.client.ui.dialogs.phaseDisplay.TargetChoiceDialog;
@@ -70,7 +69,6 @@ import megamek.common.board.Board;
 import megamek.common.board.Coords;
 import megamek.common.compute.Compute;
 import megamek.common.compute.ComputeArc;
-import megamek.common.compute.TurretFacing;
 import megamek.common.enums.AimingMode;
 import megamek.common.equipment.AmmoMounted;
 import megamek.common.equipment.AmmoType;
@@ -1465,6 +1463,7 @@ public class FiringDisplay extends AttackPhaseDisplay implements ListSelectionLi
     /**
      * Removes all current fire
      */
+    @Override
     protected void clearAttacks() {
         isStrafing = false;
         strafingCoords.clear();
@@ -2073,195 +2072,12 @@ public class FiringDisplay extends AttackPhaseDisplay implements ListSelectionLi
     }
 
     /**
-     * Flips the currently selected weapon's Directional Torso Mount (BMM p.83) between the front and rear arc. The
-     * facing is a mount offset relative to the unit's (possibly twisted) facing and is declared as a pending action
-     * alongside the attacks, so it composes with a torso twist in either order without cancelling it. Every weapon in
-     * the mount's location shares the facing and flips together. Unlike a torso twist the chosen arc persists across
-     * rounds. Ignored if the mount has been locked by damage.
+     * Refreshes the firing-phase target panel after a Directional Torso Mount arc change. See
+     * {@link AttackPhaseDisplay#flipDirectionalMount()}, which drives the shared flip logic.
      */
-    public void flipDirectionalMount() {
-        if (currentEntity() == null) {
-            logger.info("[DirTorsoMount] flip ignored - no current entity");
-            return;
-        }
-        WeaponMounted weapon = clientgui.getUnitDisplay().wPan.getSelectedWeapon();
-        int weaponNumber = clientgui.getUnitDisplay().wPan.getSelectedWeaponNum();
-        if ((weapon == null) || (weaponNumber == -1)) {
-            logger.info("[DirTorsoMount] {}: flip ignored - no weapon selected (weaponNumber={})",
-                  currentEntity().getShortName(), weaponNumber);
-            return;
-        }
-        logger.info("[DirTorsoMount] {}: flip requested - {}",
-              currentEntity().getShortName(), weapon.directionalMountDebug());
-        if (!weapon.hasDirectionalTorsoMount()) {
-            logger.info("[DirTorsoMount] {}: flip ignored - {} is not a directional mount",
-                  currentEntity().getShortName(), weapon.getName());
-            return;
-        }
-        if (weapon.hasDirectional360TorsoMount()) {
-            logger.info("[DirTorsoMount] {}: flip ignored - {} is a 360-degree turret (no front/rear to flip)",
-                  currentEntity().getShortName(), weapon.getName());
-            return;
-        }
-        if (weapon.isDirectionalMountLocked()) {
-            logger.info("[DirTorsoMount] {}: flip ignored - mount for {} is locked",
-                  currentEntity().getShortName(), weapon.getName());
-            return;
-        }
-        boolean newRear = !weapon.isDirectionalMountRear();
-        int newFacing = newRear ? 3 : 0;
-        int flippedLocation = weapon.getLocation();
-
-        // clearAttacks() below resets the secondary facing and drops every pending action; capture the declared
-        // torso twist, arms-flip and any other-location mount facings so the flip does not undo them (BMM p.83:
-        // the mount arc is declared alongside - and independently of - a torso twist).
-        int twistedFacing = currentEntity().getSecondaryFacing();
-        boolean wasTwisted = twistedFacing != currentEntity().getFacing();
-        boolean wasArmsFlipped = currentEntity().getArmsFlipped();
-        List<DirectionalMountFacingAction> otherMountFacings = pendingDirectionalMountFacings(flippedLocation);
-
-        clearAttacks();
-
-        if (wasTwisted) {
-            currentEntity().setSecondaryFacing(twistedFacing);
-            addAttack(new TorsoTwistAction(currentEntity, twistedFacing));
-        }
-        if (wasArmsFlipped) {
-            currentEntity().setArmsFlipped(true);
-            addAttack(new FlipArmsAction(currentEntity, true));
-        }
-        for (DirectionalMountFacingAction mountFacing : otherMountFacings) {
-            addAttack(mountFacing);
-        }
-
-        // The whole mount (every directional weapon in this location) shares one facing, so flip them together.
-        DirectionalTorsoMountRules.setMountFacing(currentEntity(), flippedLocation, newFacing);
-        addAttack(new DirectionalMountFacingAction(currentEntity, weaponNumber, newFacing));
-        logger.info("[DirTorsoMount] {}: {} flipped to {} (facing offset {}); queued facing action",
-              currentEntity().getShortName(), weapon.getName(), newRear ? "REAR" : "FRONT", newFacing);
+    @Override
+    protected void refreshTargetAfterMountChange() {
         updateTarget();
-        // Rebuild the weapon display and reselect the flipped weapon (mirrors the mode-change flow). refreshAll()
-        // is avoided here: it calls selectFirstWeapon() and short-circuits the unit-display rebuild when the entity
-        // object is unchanged, which is what dropped the selection and left the arc indicator stale.
-        clientgui.onAllBoardViews(boardView -> boardView.redrawEntity(currentEntity()));
-        clientgui.getUnitDisplay().wPan.displayMek(currentEntity());
-        clientgui.getUnitDisplay().wPan.selectWeapon(weapon);
-        updateDonePanel();
-        clientgui.updateFiringArc(currentEntity());
-    }
-
-    /**
-     * Collects the pending Directional Torso Mount facing actions (BMM p.83) so they can be re-added after a
-     * {@link #clearAttacks()} that is only meant to rebuild the weapon attacks (a torso twist or another mount flip).
-     * The mount arc is a persistent, separately-declared state, not a weapon attack, so it must survive the clear.
-     *
-     * @param excludedLocation a location to omit because the caller re-declares it itself, or -1 to keep all
-     *
-     * @return the pending mount facing actions to preserve
-     */
-    private List<DirectionalMountFacingAction> pendingDirectionalMountFacings(int excludedLocation) {
-        List<DirectionalMountFacingAction> mountFacings = new ArrayList<>();
-        if (currentEntity() == null) {
-            return mountFacings;
-        }
-        for (EntityAction action : attacks) {
-            if (action instanceof DirectionalMountFacingAction mountFacing) {
-                Mounted<?> mount = currentEntity().getEquipment(mountFacing.getWeaponNumber());
-                if ((mount != null) && (mount.getLocation() != excludedLocation)) {
-                    mountFacings.add(mountFacing);
-                }
-            }
-        }
-        return mountFacings;
-    }
-
-    /**
-     * Enables the flip-mount button only when the selected weapon is in a 2-point Directional Torso Mount whose arc
-     * can still be changed - i.e. it is a directional mount, is not a 360-degree turret (which has no front/rear to
-     * flip), and has not been locked by damage. Logs the decision for any directional-mount unit to aid playtesting.
-     */
-    protected void updateFlipMount() {
-        WeaponMounted weapon = clientgui.getUnitDisplay().wPan.getSelectedWeapon();
-        if ((currentEntity() == null) || (weapon == null)) {
-            setFlipMountEnabled(false);
-            return;
-        }
-        boolean canFlip = weapon.hasDirectionalTorsoMount()
-              && !weapon.hasDirectional360TorsoMount()
-              && !weapon.isDirectionalMountLocked();
-        boolean unitHasDirectionalMount =
-              currentEntity().hasQuirk(OptionsConstants.QUIRK_POS_DIRECTIONAL_TORSO_MOUNT)
-                    || currentEntity().hasQuirk(OptionsConstants.QUIRK_POS_DIRECTIONAL_TORSO_MOUNT_360)
-                    || weapon.hasDirectionalTorsoMount();
-        if (unitHasDirectionalMount) {
-            logger.info("[DirTorsoMount] {}: flip-mount button {} - {}",
-                  currentEntity().getShortName(), canFlip ? "ENABLED" : "disabled",
-                  weapon.directionalMountDebug());
-        }
-        setFlipMountEnabled(canFlip);
-    }
-
-    /**
-     * Opens the facing dialog for the selected weapon's rotatable mount - a Mek shoulder/head/quad turret, a vehicle
-     * dual turret, or a Directional Torso Mount (issues #1040, #6518) - so turrets can be aimed with an action button,
-     * not only the right-click menu.
-     */
-    public void rotateSelectedMount() {
-        Entity entity = currentEntity();
-        if (entity == null) {
-            return;
-        }
-        WeaponMounted weapon = clientgui.getUnitDisplay().wPan.getSelectedWeapon();
-        if (weapon == null) {
-            return;
-        }
-        if (weapon.hasDirectionalTorsoMount() && (entity instanceof Mek directionalMek)) {
-            new TurretFacingDialog(clientgui.getFrame(), directionalMek, weapon, clientgui, true).setVisible(true);
-        } else if (weapon.isMekTurretMounted() && (entity instanceof Mek turretMek)) {
-            Mounted<?> turretItem = findMekTurretItem(turretMek, weapon.getLocation());
-            if (turretItem != null) {
-                new TurretFacingDialog(clientgui.getFrame(), turretMek, turretItem, clientgui).setVisible(true);
-            }
-        } else if ((entity instanceof Tank tank) && (weapon.getLocation() == tank.getLocTurret2())) {
-            new TurretFacingDialog(clientgui.getFrame(), tank, clientgui).setVisible(true);
-        }
-    }
-
-    /**
-     * @param mek      the unit
-     * @param location the location holding a Mek turret
-     *
-     * @return the turret {@link MiscType} equipment item in the location (shoulder/head/quad turret), or {@code null}
-     */
-    private Mounted<?> findMekTurretItem(Mek mek, int location) {
-        for (Mounted<?> misc : mek.getMisc()) {
-            if ((misc.getLocation() == location) && (misc.getType().hasFlag(MiscType.F_SHOULDER_TURRET)
-                  || misc.getType().hasFlag(MiscType.F_HEAD_TURRET)
-                  || misc.getType().hasFlag(MiscType.F_QUAD_TURRET))) {
-                return misc;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Enables the rotate-turret button when the selected weapon sits on a rotatable mount (a Mek turret, a vehicle dual
-     * turret, or a Directional Torso Mount) whose facing can still be changed.
-     */
-    protected void updateRotateTurret() {
-        WeaponMounted weapon = clientgui.getUnitDisplay().wPan.getSelectedWeapon();
-        if ((currentEntity() == null) || (weapon == null)) {
-            setRotateTurretEnabled(false);
-            return;
-        }
-        // A 2-point Directional Torso Mount is flipped front/rear with the Flip Mount button, so it does not use
-        // the rotate dialog; only the 3-point 360 quad turret (and Mek/vehicle turrets) use the rotate button.
-        boolean isTwoPointDirectionalMount = weapon.hasDirectionalTorsoMount() && !weapon.hasDirectional360TorsoMount();
-        boolean isLockedDirectionalMount = weapon.hasDirectionalTorsoMount() && weapon.isDirectionalMountLocked();
-        boolean canRotate = TurretFacing.isRotatable(currentEntity(), weapon)
-              && !isTwoPointDirectionalMount
-              && !isLockedDirectionalMount;
-        setRotateTurretEnabled(canRotate);
     }
 
     protected void updateSearchlight() {
@@ -2483,11 +2299,13 @@ public class FiringDisplay extends AttackPhaseDisplay implements ListSelectionLi
         clientgui.getMenuBar().setEnabled(FiringCommand.FIRE_FLIP_ARMS.getCmd(), enabled);
     }
 
+    @Override
     protected void setFlipMountEnabled(boolean enabled) {
         buttons.get(FiringCommand.FIRE_FLIP_MOUNT).setEnabled(enabled);
         clientgui.getMenuBar().setEnabled(FiringCommand.FIRE_FLIP_MOUNT.getCmd(), enabled);
     }
 
+    @Override
     protected void setRotateTurretEnabled(boolean enabled) {
         buttons.get(FiringCommand.FIRE_ROTATE_TURRET).setEnabled(enabled);
         clientgui.getMenuBar().setEnabled(FiringCommand.FIRE_ROTATE_TURRET.getCmd(), enabled);
